+++ app/src/main/java/org/simple/clinic/summary/addphone/AddPhoneNumberUiRender.kt
@@ -1,0 +1,9 @@
+package org.simple.clinic.summary.addphone
+
+import org.simple.clinic.mobius.ViewRenderer
+
+class AddPhoneNumberUiRender(private val ui: AddPhoneNumberUi) : ViewRenderer<AddPhoneNumberModel> {
+  override fun render(model: AddPhoneNumberModel) {
+
+  }
+}