+++ app/src/main/java/org/simple/clinic/summary/addphone/UiActions.kt
@@ -1,0 +1,3 @@
+package org.simple.clinic.summary.addphone
+
+interface UiActions