+++ app/src/test/java/org/simple/clinic/summary/addphone/AddPhoneNumberDialogControllerTest.kt
@@ -6,18 +6,23 @@ import com.nhaarman.mockitokotlin2.never
 import com.nhaarman.mockitokotlin2.verify
 import com.nhaarman.mockitokotlin2.verifyNoMoreInteractions
 import com.nhaarman.mockitokotlin2.whenever
+import com.spotify.mobius.Init
 import io.reactivex.Completable
 import io.reactivex.disposables.Disposable
+import io.reactivex.rxkotlin.ofType
 import io.reactivex.subjects.PublishSubject
 import org.junit.After
 import org.junit.Rule
 import org.junit.Test
+import org.simple.clinic.mobius.first
 import org.simple.clinic.patient.PatientRepository
 import org.simple.clinic.patient.PhoneNumberDetails
 import org.simple.clinic.registration.phone.LengthBasedNumberValidator
 import org.simple.clinic.util.RxErrorsRule
+import org.simple.clinic.util.scheduler.TestSchedulersProvider
 import org.simple.clinic.uuid.FakeUuidGenerator
 import org.simple.clinic.widgets.UiEvent
+import org.simple.mobius.migration.MobiusTestFixture
 import java.util.UUID
 
 class AddPhoneNumberDialogControllerTest {
@@ -26,7 +31,7 @@ class AddPhoneNumberDialogControllerTest {
   val rxErrorsRule = RxErrorsRule()
 
   private val uiEvents = PublishSubject.create<UiEvent>()
-  private val dialog = mock<AddPhoneNumberDialog>()
+  private val ui = mock<AddPhoneNumberUi>()
   private val repository = mock<PatientRepository>()
 
   private val validator = LengthBasedNumberValidator(
@@ -40,10 +45,12 @@ class AddPhoneNumberDialogControllerTest {
   private val generatedPhoneUuid = UUID.fromString("f94bd99b-b182-4138-8e77-d91908b7ada5")
 
   private lateinit var controllerSubscription: Disposable
+  private lateinit var testFixture: MobiusTestFixture<AddPhoneNumberModel, AddPhoneNumberEvent, AddPhoneNumberEffect>
 
   @After
   fun tearDown() {
     controllerSubscription.dispose()
+    testFixture.dispose()
   }
 
   @Test
@@ -71,8 +78,8 @@ class AddPhoneNumberDialogControllerTest {
         active = true
     )
     verifyNoMoreInteractions(repository)
-    verify(dialog).dismiss()
-    verifyNoMoreInteractions(dialog)
+    verify(ui).closeDialog()
+    verifyNoMoreInteractions(ui)
   }
 
   @Test
@@ -88,8 +95,8 @@ class AddPhoneNumberDialogControllerTest {
     verify(repository, never()).createPhoneNumberForPatient(any(), any(), any(), any())
     verifyNoMoreInteractions(repository)
 
-    verify(dialog).showPhoneNumberTooShortError()
-    verifyNoMoreInteractions(dialog)
+    verify(ui).showPhoneNumberBlank()
+    verifyNoMoreInteractions(ui)
   }
 
   @Test
@@ -105,8 +112,8 @@ class AddPhoneNumberDialogControllerTest {
     verify(repository, never()).createPhoneNumberForPatient(any(), any(), any(), any())
     verifyNoMoreInteractions(repository)
 
-    verify(dialog).showPhoneNumberTooShortError()
-    verifyNoMoreInteractions(dialog)
+    verify(ui).showPhoneNumberTooShortError(6)
+    verifyNoMoreInteractions(ui)
   }
 
   @Test
@@ -122,8 +129,8 @@ class AddPhoneNumberDialogControllerTest {
     verify(repository, never()).createPhoneNumberForPatient(any(), any(), any(), any())
     verifyNoMoreInteractions(repository)
 
-    verify(dialog).showPhoneNumberTooLongError()
-    verifyNoMoreInteractions(dialog)
+    verify(ui).showPhoneNumberTooLongError(12)
+    verifyNoMoreInteractions(ui)
   }
 
   private fun setupController() {
@@ -134,8 +141,25 @@ class AddPhoneNumberDialogControllerTest {
         patientUuid
     )
 
+    val effectHandler = AddPhoneNumberEffectHandler(
+        schedulersProvider = TestSchedulersProvider.trampoline(),
+        uiActions = ui
+    )
+    val uiRenderer = AddPhoneNumberUiRender(ui)
+
     controllerSubscription = uiEvents
         .compose(controller)
-        .subscribe { uiChange -> uiChange(dialog) }
+        .subscribe { uiChange -> uiChange(ui) }
+
+    testFixture = MobiusTestFixture(
+        events = uiEvents.ofType(),
+        defaultModel = AddPhoneNumberModel.create(),
+        update = AddPhoneNumberUpdate(),
+        effectHandler = effectHandler.build(),
+        modelUpdateListener = uiRenderer::render,
+        init = Init { first(it) }
+    )
+
+    testFixture.start()
   }
 }